+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000214.rb
@@ -1,6 +1,6 @@
 control 'PHTN-40-000214' do
   title 'The Photon operating system must configure Secure Shell (SSH) to disallow Kerberos authentication.'
-  desc  "If Kerberos is enabled through Secure Shell (SSH), sshd provides a means of access to the system's Kerberos implementation. Vulnerabilities in the system's Kerberos implementation may then be subject to exploitation. To reduce the attack surface of the system, the Kerberos authentication mechanism within SSH must be disabled."
+  desc  "If Kerberos is enabled through SSH, sshd provides a means of access to the system's Kerberos implementation. Vulnerabilities in the system's Kerberos implementation may then be subject to exploitation. To reduce the attack surface of the system, the Kerberos authentication mechanism within SSH must be disabled."
   desc  'rationale', ''
   desc  'check', "
     At the command line, run the following command to verify the running configuration of sshd: