+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000231.rb
@@ -26,9 +26,9 @@ control 'PHTN-40-000231' do
 
     At the command line, run the following command to load the new configuration:
 
-    # /sbin/sysctl --load
+    # /sbin/sysctl --load /etc/sysctl.d/zz-stig-hardening.conf
 
-    Note: If the file zz-stig-hardening.conf doesn't exist it must be created.
+    Note: If the file zz-stig-hardening.conf does not exist it must be created.
   "
   impact 0.5
   tag severity: 'medium'
@@ -39,7 +39,14 @@ control 'PHTN-40-000231' do
   tag cci: ['CCI-000366']
   tag nist: ['CM-6 b']
 
-  describe kernel_parameter('net.ipv4.ip_forward') do
-    its('value') { should cmp 0 }
+  if input('containerHost')
+    impact 0.0
+    describe 'Container host flag set to true so this control is N/A...skipping...' do
+      skip 'Container host flag set to true so this control is N/A...skipping...'
+    end
+  else
+    describe kernel_parameter('net.ipv4.ip_forward') do
+      its('value') { should cmp 0 }
+    end
   end
 end