+++ photon/4.0/inspec/vmware-photon-4.0-stig-baseline/controls/PHTN-40-000238.rb
@@ -28,7 +28,7 @@ control 'PHTN-40-000238' do
 
     # /sbin/augenrules --load
 
-    Note: A \"audit.STIG.rules\" file is provided with this guidance for placement in \"/etc/audit/rules.d\" that contains all rules needed for auditd.
+    Note: An \"audit.STIG.rules\" file is provided with this guidance for placement in \"/etc/audit/rules.d\" that contains all rules needed for auditd.
 
     Note: An older \"audit.STIG.rules\" may exist and may reference older \"GEN\" SRG IDs. This file can be removed and replaced as necessary with an updated one.
   "