+++ photon/5.0/inspec/vmware-photon-5.0-stig-baseline/controls/PHTN-50-000107.rb
@@ -5,7 +5,7 @@ control 'PHTN-50-000107' do
   desc  'check', "
     At the command line, run the following command to output a list of files with setuid/setgid configured and their corresponding audit rules:
 
-    # for file in $(find / -xdev -path /var/lib/containerd -prune -o \\( -perm -4000 -o -perm -2000 \\) -type f -print | sort); do echo \"Found file with setuid/setgid configured: $file\";rule=\"$(auditctl -l | grep \"$file \")\";echo \"Audit Rule Result: $rule\";echo \"\"; done
+    # for file in $(find / -xdev -path /var/lib/containerd -prune -o -path /var/lib/docker -prune -o \\( -perm -4000 -o -perm -2000 \\) -type f -print | sort); do echo \"Found file with setuid/setgid configured: $file\";rule=\"$(auditctl -l | grep \"$file \")\";echo \"Audit Rule Result: $rule\";echo \"\"; done
 
     Example output:
 
@@ -48,7 +48,7 @@ control 'PHTN-50-000107' do
   tag cci: ['CCI-000172', 'CCI-001404', 'CCI-002234']
   tag nist: ['AC-2 (4)', 'AC-6 (9)', 'AU-12 c']
 
-  results = command('find / -xdev -path /var/lib/containerd -prune -o \( -perm -4000 -type f -o -perm -2000 \) -type f -print').stdout.split("\n")
+  results = command('find / -xdev -path /var/lib/containerd -prune -o -path /var/lib/docker -prune -o \( -perm -4000 -type f -o -perm -2000 \) -type f -print').stdout.split("\n")
   if !results.empty?
     results.each do |path|
       describe auditd do